requirements.txt
streamlit==1.34.0
langchain==0.1.5
langchain-community==0.0.19
langchain-ollama==0.0.1
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
networkx==3.1
scikit-learn==1.3.2
rank-bm25==0.2.2
torch==2.1.0
docx2txt==0.8
